Move gtsam extension code to new wave_gtsam module
wave_gtsam now depends only on gtsam, so its useful classes can be used without
wave_vision or ceres.  wave_optimization depends on ceres and wave_vision, but
it is mostly useless for now.
