+++ wave_gtsam/include/wave/gtsam/decaying_bias.hpp
