+++ wave_gtsam/include/wave/gtsam/hand_eye.hpp
