+++ wave_gtsam/include/wave/gtsam/impl/pose_prior_impl.hpp
