+++ wave_gtsam/src/gps_factor_with_bias.cpp
@@ -1,4 +1,4 @@
-#include "wave/optimization/gtsam/gps_factor_with_bias.hpp"
+#include "wave/gtsam/gps_factor_with_bias.hpp"
 
 namespace wave {
