+++ wave_gtsam/tests/gtsam/gps_factor_with_bias_test.cpp
@@ -5,9 +5,9 @@
 #include <gtsam/nonlinear/GaussNewtonOptimizer.h>
 #include <gtsam/slam/PriorFactor.h>
 #include <gtsam/slam/BetweenFactor.h>
-#include "wave/optimization/gtsam/gps_factor_with_bias.hpp"
-#include "wave/optimization/gtsam/gps_factor_with_bias_general.hpp"
-#include "wave/optimization/gtsam/pose_vel_bias.hpp"
+#include "wave/gtsam/gps_factor_with_bias.hpp"
+#include "wave/gtsam/gps_factor_with_bias_general.hpp"
+#include "wave/gtsam/pose_vel_bias.hpp"
 #include "wave/wave_test.hpp"
 
 namespace wave {
